experiments/h1_split_mnist/requirements.txt
torch>=2.4
torchvision>=0.19
avalanche-lib>=0.6
numpy>=1.26
pandas>=2.0
